+++ src/Entities/Order.php
@@ -11,6 +11,7 @@ use GingerPluginSdk\Helpers\HelperTrait;
 use GingerPluginSdk\Helpers\MultiFieldsEntityTrait;
 use GingerPluginSdk\Helpers\SingleFieldTrait;
 use GingerPluginSdk\Interfaces\MultiFieldsEntityInterface;
+use GingerPluginSdk\Properties\Amount;
 use GingerPluginSdk\Properties\Currency;
 use GingerPluginSdk\Properties\Status;
 
@@ -22,7 +23,6 @@ class Order implements MultiFieldsEntityInterface
     use SingleFieldTrait;
 
     private BaseField $merchantOrderId;
-    private BaseField $amount;
     private BaseField $webhookUrl;
     private BaseField $returnUrl;
     private BaseField $description;
@@ -36,73 +36,24 @@ class Order implements MultiFieldsEntityInterface
 
     public function __construct(
         private Currency     $currency,
-        float|int            $amount,
+        private Amount       $amount,
         private Transactions $transactions,
         private Customer     $customer,
         private ?OrderLines  $orderLines = null,
-        ?string              $returnUrl = null,
-        ?string              $webhookUrl = null,
-        ?string              $merchantOrderId = null,
-        ?string              $description = null,
         private ?Extra       $extra = null,
         private ?Client      $client = null,
-        ?string              $created = null,
         private ?Flags       $flags = null,
         ?string              $id = null,
-        ?string              $lastTransactionAdded = null,
-        ?string              $merchantId = null,
-        ?string              $modified = null,
-        ?string              $projectId = null,
         private ?Status      $status = null,
-        ?string              $completed = null,
+        mixed                ...$additionalProperties
     )
     {
-        $this->amount = $this->createSimpleField(
-            propertyName: 'amount',
-            value: $this->calculateValueInCents($amount)
-        );
-
-        if ($completed) $this->createFieldInDateTimeISO8601(
-            propertyName: 'completed',
-            value: $completed
-        );
-
-        if ($created) $this->created = $this->createFieldInDateTimeISO8601(
-            propertyName: 'created',
-            value: $created
-        );
-
         if ($id) $this->id = $this->createSimpleField(
             propertyName: 'id',
             value: $id
         );
 
-        if ($lastTransactionAdded) $this->lastTransactionAdded = $this->createFieldInDateTimeISO8601(
-            propertyName: 'last_transaction_added',
-            value: $lastTransactionAdded
-        );
-
-        if ($merchantId) $this->merchantId = $this->createSimpleField(
-            propertyName: 'merchant_id',
-            value: $merchantId
-        );
-
-        if ($modified) $this->modified = $this->createFieldInDateTimeISO8601(
-            propertyName: 'modified',
-            value: $modified
-        );
-
-        if ($projectId) $this->projectId = $this->createSimpleField(
-            propertyName: 'project_id',
-            value: $projectId
-        );
-
-        $this->setWebhookUrl($webhookUrl)
-            ->setMerchantOrderId($merchantOrderId)
-            ->setReturnUrl($returnUrl)
-            ->setDescription($description)
-            ->setExtra($extra)
-            ->setClient($client);
+        if ($additionalProperties) $this->filterAdditionalProperties($additionalProperties);
     }
 
     public function getId(): string|null
@@ -132,7 +83,7 @@ class Order implements MultiFieldsEntityInterface
 
     public function getOrderLines(): ?OrderLines
     {
-        return $this->orderLines?->get();
+        return $this->orderLines;
     }
 
     public function getReturnUrl(): string
@@ -163,83 +114,5 @@ class Order implements MultiFieldsEntityInterface
     public function getCurrentTransaction(): Transaction
     {
         return $this->transactions->get();
-    }
-
-    public function setClient(?Client $client): Order
-    {
-        $this->client = $client;
-        return $this;
-    }
-
-    /**
-     * @param string|null $id_string
-     * @return $this
-     */
-    public function setMerchantOrderId(?string $id_string): Order
-    {
-        $this->merchantOrderId = $this->createSimpleField(
-            propertyName: 'merchant_order_id',
-            value: $id_string
-        );
-        return $this;
-    }
-
-    /**
-     * @param string|null $url
-     * @return $this
-     */
-    public function setWebhookUrl(?string $url): Order
-    {
-        $this->webhookUrl = $this->createSimpleField(
-            propertyName: 'webhook_url',
-            value: $url
-        );
-        return $this;
-    }
-
-    /**
-     * @param string|null $url
-     * @return $this
-     */
-    public function setReturnUrl(?string $url): Order
-    {
-        $this->returnUrl = $this->createSimpleField(
-            propertyName: 'return_url',
-            value: $url
-        );
-        return $this;
-    }
-
-    /**
-     * @param OrderLines $lines
-     * @return $this
-     */
-    public function setOrderLines(OrderLines $lines): Order
-    {
-        $this->orderLines = $lines;
-        return $this;
-    }
-
-    /**
-     * @param \GingerPluginSdk\Entities\Extra|null $extra
-     * @return $this
-     */
-    public function setExtra(?Extra $extra): Order
-    {
-        $this->extra = $extra;
-        return $this;
-    }
-
-    /**
-     * @param string|null $description
-     * @return $this
-     */
-    public function setDescription(?string $description): Order
-    {
-        $this->description = $this->createSimpleField(
-            propertyName: 'description',
-            value: $description
-        );
-        return $this;
     }
 }