+++ src/Tests/FromArrayTest.php
@@ -14,6 +14,7 @@ use GingerPluginSdk\Entities\Line;
 use GingerPluginSdk\Entities\Order;
 use GingerPluginSdk\Entities\PaymentMethodDetails;
 use GingerPluginSdk\Entities\Transaction;
+use GingerPluginSdk\Properties\Amount;
 use GingerPluginSdk\Properties\ClientOptions;
 use GingerPluginSdk\Properties\Country;
 use GingerPluginSdk\Properties\Birthdate;
@@ -86,7 +87,7 @@ class FromArrayTest extends TestCase
         $_SERVER["HTTP_USER_AGENT"] = "PHPUnit Tests";
         $expected_order = new Order(
             currency: new Currency('EUR'),
-            amount: 500,
+            amount: new Amount(500),
             transactions: new Transactions(
                 new Transaction(
                     paymentMethod: 'ideal',
@@ -139,7 +140,7 @@ class FromArrayTest extends TestCase
                     merchantOrderLineId: "5",
                     name: 'Milk',
                     quantity: 1,
-                    amount: 1.00,
+                    amount: new Amount(1.00),
                     vatPercentage: 50,
                     currency: new Currency(
                         'EUR'